// client/src/types/ObjectDestroy.ts
export default interface DestructionObject {
    title: string;
    position: any[];
    postName: string;
    address: string;
    typeInfrastructure: string;
    area: number;
    imgPath: string;
    description: string;
    percentageOfDestruction: string;
    dateOfDestruction: string;
    dateOfRecovery: string;
    typeDestruction: string;
    countVictims: number;
    whatDestroyed: string;
    areaName: string;
    neighborhood: string,
    stateDestruction: string,
    _id: string | undefined; 
}